+++ src/windows/mod.rs
@@ -18,6 +18,7 @@
 mod ipfp;
 #[cfg(feature = "registry")]
 mod registry;
+mod sys;
 
 pub(crate) use ipfp::fill as fill_ipfp;
 #[cfg(feature = "registry")]